Slot module: applied-plans rename + MergePolicy extraction + injectPartial rename (internal, behavior-preserving)
Renames (Slot.md Step 2): lifecycle/browser-plans.ts -> lifecycle/applied-plans.ts;
class AppliedBrowserPlans -> AppliedPlans; singleton appliedBrowserPlans -> appliedPlans;
execution/inject.ts export injectHtml -> injectPartial. All importers/refs updated
(boot.ts, root.ts, execute.ts call site, and the slot/boot/inject/gather/validation tests).

Extraction (Slot.md Step 3): new lifecycle/merge-policy.ts exports the MergePolicy
object (composeBootPlanInto + composeSlotPlanInto) plus emptyPlan/snapshotPlan, moved
out of applied-plans.ts. Pure move; recompose + composeInitialPlans now delegate to
MergePolicy. No logic change.

Step 4 (recompose fresh-document) PARKED: building a fresh PlanDocument on recompose
instead of in-place resetPlanDocument changes object identity. Two existing gather
lifecycle assertions (all-registered-inputs-lifecycle.test.ts) capture the registered
active-plan reference before a slot load and rely on in-place mutation; a fresh-doc
recompose makes that reference stale and fails those assertions. Behavior must be
preserved without changing any test assertion, so recompose keeps in-place reuse.

Gate: dotnet build 0 errors (both TFMs); npm run typecheck 0; vitest 192/192 green
(same assertions); 0 stale injectHtml/AppliedBrowserPlans/browser-plans refs in code;
plan.ts byte-stable.
