+++ Alis.Reactive.Assets/runtime/__tests__/plan-lifecycle/boot-plan-composition.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { composeInitialPlans } from "../../lifecycle/browser-plans";
+import { composeInitialPlans } from "../../lifecycle/applied-plans";
 import type { ComponentObject } from "../../types";
 import {
   behavior,
