+++ Alis.Reactive.Assets/runtime/lifecycle/boot.ts
@@ -1,6 +1,6 @@
 // Boot — PlanDocument composition: boot, partial slot load/unload, reset.
 // Single responsibility: wire behaviors and validation for active plans.
-// Delegates plan composition state to browser-plans.ts.
+// Delegates plan composition state to applied-plans.ts.
 
 import type { PlanDocument, Behavior } from "../types";
 import { setLevel } from "../core/trace";
@@ -12,9 +12,9 @@ import { findSummaryElement, clearSummary, hideSummaryDiv } from "../validation/
 import { resetNativeActionLinksForTests } from "../components/native/native-action-link";
 import { resetPluginCatalogForTests } from "../core/plugin-catalog";
 import {
-  appliedBrowserPlans,
+  appliedPlans,
   type BrowserPlanWiring,
-} from "./browser-plans";
+} from "./applied-plans";
 
 const log = scope("boot");
 const BOOTED_ATTR = "alisBooted";
@@ -40,7 +40,7 @@ export function boot(plan: PlanDocument): void {
   wireBehaviors(plan.behaviors, plan, bootAbort.signal);
 
   setActivePlan(plan);
-  appliedBrowserPlans.register(plan);
+  appliedPlans.register(plan);
   markReactiveBooted(plan);
   log.info("booted", { planId: plan.planId });
 }
@@ -72,7 +72,7 @@ function wireContainerValidation(plan: PlanDocument, signal?: AbortSignal): void
 }
 
 export function loadPartialSlot(slotId: string, incoming: PlanDocument[]): void {
-  const affectedPlanIds = appliedBrowserPlans.loadPartialSlot(slotId, incoming, browserPlanWiring());
+  const affectedPlanIds = appliedPlans.loadPartialSlot(slotId, incoming, browserPlanWiring());
 
   for (const planId of affectedPlanIds) {
     clearSummaryForPlan(planId);
@@ -89,7 +89,7 @@ export function loadPartialSlot(slotId: string, incoming: PlanDocument[]): void 
 }
 
 export function unloadPartialSlot(slotId: string): void {
-  const affectedPlanIds = appliedBrowserPlans.unloadPartialSlot(slotId);
+  const affectedPlanIds = appliedPlans.unloadPartialSlot(slotId);
 
   for (const planId of affectedPlanIds) {
     clearSummaryForPlan(planId);
@@ -102,7 +102,7 @@ export function unloadPartialSlot(slotId: string): void {
 }
 
 export function getBootedPlan(planId: string): PlanDocument | undefined {
-  return appliedBrowserPlans.get(planId);
+  return appliedPlans.get(planId);
 }
 
 export function resetBootStateForTests(): void {
@@ -124,7 +124,7 @@ function browserPlanWiring(): BrowserPlanWiring {
 
 function resetRuntimeSingletonsForTests(): void {
   resetActivePlanForTests();
-  appliedBrowserPlans.reset();
+  appliedPlans.reset();
   resetLiveClearForTests();
   resetNativeActionLinksForTests();
   resetPluginCatalogForTests();
